# app/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: API de Node MongoDB para INESDI
  version: 1.0.0
  description: Esta es una aplicación API simple de CRUD hecha con Express y documentada con Swagger
  contact:
    name: INESDI
    url: https://www.inesdi.com
servers:
  - url: 'http://localhost:3000'
    description: Servidor de Desarrollo
paths:
  /api/items:
    get:
      summary: Obtener todos los ítems
      responses:
        '200':
          description: Lista de ítems
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
    put:
      summary: Actualizar un ítem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Item'
      responses:
        '200':
          description: El ítem fue actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
    delete:
      summary: Eliminar ítems basados en parámetros de consulta
      parameters:
        - in: query
          name: first_name
          schema:
            type: string
          description: Nombre del individuo a eliminar
      responses:
        '200':
          description: Cantidad de ítems eliminados
          content:
            application/json:
              schema:
                type: object
                properties:
                  deletedCount:
                    type: integer
        '204':
          description: No se encontraron ítems para eliminar
  /api/items/search:
    get:
      summary: Buscar ítems basados en parámetros de consulta
      parameters:
        - in: query
          name: first_name
          schema:
            type: string
          description: Nombre del individuo a coincidir
      responses:
        '200':
          description: Un arreglo de ítems que coinciden con la consulta
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
components:
  schemas:
    Item:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
        gender:
          type: string
        address:
          type: object
          properties:
            city:
              type: string
            state:
              type: string
            country:
              type: string
            country_code:
              type: string
        card:
          type: object
          properties:
            card_number:
              type: string
            card_type:
              type: string
            currency_code:
              type: string
            balance:
              type: string
        married_status:
          type: boolean